requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
MetaTrader5>=5.0.45
tensorflow>=2.13.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
tables>=3.8.0
yfinance>=0.2.0

